pyproject.toml
[tool]
[tool.poetry]
name = "poetry-pyspark-devcontainer"
version = "0.0.0"
homepage = "https://github.com/ghjklw/poetry-pyspark-devcontainer"
repository = "https://github.com/ghjklw/poetry-pyspark-devcontainer"
documentation = "https://github.com/ghjklw/poetry-pyspark-devcontainer/README.md"
description = "PySpark project using Poetry and DevContainer."
authors = [
    "Malo Jaffré <malo.jaffre@bouvet.no>",
]
readme = "README.md"
packages = [
    { include = "poetry_pyspark_devcontainer", from = "src" },
    { include = "tests", format = "sdist" },
]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
pyspark = "~3.3.0"
delta-spark = "2.2.0"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.2.0"
black  = "^23.3.0"
pytest = "^7.3.1"
chispa = "^0.9.2"
ruff = "^0.0.290"
ipykernel = "^6.24.0"

[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 99
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.coverage.paths]
source = ["src", "*/site-packages"]
tests = ["tests"]

[tool.coverage.run]
branch = true
source = ["src"]

[tool.coverage.report]
show_missing = true
include_namespace_packages = true

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.pyright]
include = ["tests", "src"]
ignore = ["tests", "*.ipynb"]
typeCheckingMode = "strict"
reportMissingImports = true
venvPath = "."
venv = ".venv"

[tool.ruff]
select = ["ALL"]
ignore = ["I001"]
line-length = 99
target-version = "py310"

[tool.ruff.pydocstyle]
convention = "google"
